Added for negative horizontal-, vertical offsett and scale
By setting the min value for horizontal_offset and vertical_offset, the offset can also be used in negative directions. This is usefull if you need an offset in the other direction. (For example because your toolhead/laser is on the left from the carriages head.) The negative gcode scaling factor allows for a decrease in size.